Forgot what I did again.
All I know for sure is that I just did a reformat lol.

+++ src/main/java/me/earth/phobos/mixin/mixins/MixinItemRenderer.java
@@ -15,53 +15,59 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(value={ItemRenderer.class})
-public abstract class MixinItemRenderer {
+@Mixin(value = {ItemRenderer.class})
+public abstract
+class MixinItemRenderer {
     private boolean injection = true;
 
     @Shadow
-    public abstract void renderItemInFirstPerson(AbstractClientPlayer var1, float var2, float var3, EnumHand var4, float var5, ItemStack var6, float var7);
+    public abstract
+    void renderItemInFirstPerson ( AbstractClientPlayer var1 , float var2 , float var3 , EnumHand var4 , float var5 , ItemStack var6 , float var7 );
 
-    @Inject(method={"renderItemInFirstPerson(Lnet/minecraft/client/entity/AbstractClientPlayer;FFLnet/minecraft/util/EnumHand;FLnet/minecraft/item/ItemStack;F)V"}, at={@At(value="HEAD")}, cancellable=true)
-    public void renderItemInFirstPersonHook(AbstractClientPlayer player, float p_187457_2_, float p_187457_3_, EnumHand hand, float p_187457_5_, ItemStack stack, float p_187457_7_, CallbackInfo info) {
-        if (this.injection) {
-            info.cancel();
-            SmallShield offset = SmallShield.getINSTANCE();
+    @Inject(method = {"renderItemInFirstPerson(Lnet/minecraft/client/entity/AbstractClientPlayer;FFLnet/minecraft/util/EnumHand;FLnet/minecraft/item/ItemStack;F)V"}, at = {@At(value = "HEAD")}, cancellable = true)
+    public
+    void renderItemInFirstPersonHook ( AbstractClientPlayer player , float p_187457_2_ , float p_187457_3_ , EnumHand hand , float p_187457_5_ , ItemStack stack , float p_187457_7_ , CallbackInfo info ) {
+        if ( this.injection ) {
+            info.cancel ( );
+            SmallShield offset = SmallShield.getINSTANCE ( );
             float xOffset = 0.0f;
             float yOffset = 0.0f;
             this.injection = false;
-            if (hand == EnumHand.MAIN_HAND) {
-                if (offset.isOn() && player.getHeldItemMainhand() != ItemStack.EMPTY) {
-                    xOffset = offset.mainX.getValue().floatValue();
-                    yOffset = offset.mainY.getValue().floatValue();
+            if ( hand == EnumHand.MAIN_HAND ) {
+                if ( offset.isOn ( ) && player.getHeldItemMainhand ( ) != ItemStack.EMPTY ) {
+                    xOffset = offset.mainX.getValue ( ).floatValue ( );
+                    yOffset = offset.mainY.getValue ( ).floatValue ( );
                 }
-            } else if (!offset.normalOffset.getValue().booleanValue() && offset.isOn() && player.getHeldItemOffhand() != ItemStack.EMPTY) {
-                xOffset = offset.offX.getValue().floatValue();
-                yOffset = offset.offY.getValue().floatValue();
+            } else if ( ! offset.normalOffset.getValue ( ).booleanValue ( ) && offset.isOn ( ) && player.getHeldItemOffhand ( ) != ItemStack.EMPTY ) {
+                xOffset = offset.offX.getValue ( ).floatValue ( );
+                yOffset = offset.offY.getValue ( ).floatValue ( );
             }
-            this.renderItemInFirstPerson(player, p_187457_2_, p_187457_3_, hand, p_187457_5_ + xOffset, stack, p_187457_7_ + yOffset);
+            this.renderItemInFirstPerson ( player , p_187457_2_ , p_187457_3_ , hand , p_187457_5_ + xOffset , stack , p_187457_7_ + yOffset );
             this.injection = true;
         }
     }
 
-    @Redirect(method={"renderArmFirstPerson"}, at=@At(value="INVOKE", target="Lnet/minecraft/client/renderer/GlStateManager;translate(FFF)V", ordinal=0))
-    public void translateHook(float x, float y, float z) {
-        SmallShield offset = SmallShield.getINSTANCE();
-        boolean shiftPos = Minecraft.getMinecraft().player != null && Minecraft.getMinecraft().player.getHeldItemMainhand() != ItemStack.EMPTY && offset.isOn();
-        GlStateManager.translate((float)(x + (shiftPos ? offset.mainX.getValue().floatValue() : 0.0f)), (float)(y + (shiftPos ? offset.mainY.getValue().floatValue() : 0.0f)), (float)z);
+    @Redirect(method = {"renderArmFirstPerson"}, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/GlStateManager;translate(FFF)V", ordinal = 0))
+    public
+    void translateHook ( float x , float y , float z ) {
+        SmallShield offset = SmallShield.getINSTANCE ( );
+        boolean shiftPos = Minecraft.getMinecraft ( ).player != null && Minecraft.getMinecraft ( ).player.getHeldItemMainhand ( ) != ItemStack.EMPTY && offset.isOn ( );
+        GlStateManager.translate ( x + ( shiftPos ? offset.mainX.getValue ( ).floatValue ( ) : 0.0f ) , y + ( shiftPos ? offset.mainY.getValue ( ).floatValue ( ) : 0.0f ) , z );
     }
 
-    @Inject(method={"renderFireInFirstPerson"}, at={@At(value="HEAD")}, cancellable=true)
-    public void renderFireInFirstPersonHook(CallbackInfo info) {
-        if (NoRender.getInstance().isOn() && NoRender.getInstance().fire.getValue().booleanValue()) {
-            info.cancel();
+    @Inject(method = {"renderFireInFirstPerson"}, at = {@At(value = "HEAD")}, cancellable = true)
+    public
+    void renderFireInFirstPersonHook ( CallbackInfo info ) {
+        if ( NoRender.getInstance ( ).isOn ( ) && NoRender.getInstance ( ).fire.getValue ( ).booleanValue ( ) ) {
+            info.cancel ( );
         }
     }
 
-    @Inject(method={"renderSuffocationOverlay"}, at={@At(value="HEAD")}, cancellable=true)
-    public void renderSuffocationOverlay(CallbackInfo ci) {
-        if (NoRender.getInstance().isOn() && NoRender.getInstance().blocks.getValue().booleanValue()) {
-            ci.cancel();
+    @Inject(method = {"renderSuffocationOverlay"}, at = {@At(value = "HEAD")}, cancellable = true)
+    public
+    void renderSuffocationOverlay ( CallbackInfo ci ) {
+        if ( NoRender.getInstance ( ).isOn ( ) && NoRender.getInstance ( ).blocks.getValue ( ).booleanValue ( ) ) {
+            ci.cancel ( );
         }
     }
 }
